pyproject.toml
[project]
name = "ai-story-twitter-bot"
version = "0.1.0"
description = "An AI-powered Twitter bot that generates and posts solarpunk micro-stories with accompanying AI-generated images"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "google-cloud-aiplatform>=1.42.1",
    "vertexai>=0.0.1",
    "tweepy>=4.14.0",
    "requests-oauthlib>=1.3.1",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "google-api-python-client>=2.108.0",
    "PyYAML>=6.0.1",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "sentry-sdk>=1.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "responses>=0.24.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]
asyncio_mode = "auto"
